Remove gcbool.h, change BOOL to bool
HUGE breaking API change, but that's what we're going for with libogc-mod, right? :wink:

This is only half the battle, though. Many functions that return boolean values are still 32-bit integers or other types. I'll have to go through those for a separate commit.

Also removed NULL redefinition, it's unnecessary.