Remove ASNDLIB
All of my focus is on AESND (because it's superior to ASND in every way), so ASND will just be unmaintained.

Because of this, I decided to remove it from libogc-mod. Since I'm currently reworking the API, there's no point in keeping it for compatibility, and therefore no point in keeping it at all.

Gone, but not forgotten. Long live ASND!